major refactor to split all classes out to separate files
